// src/api/generated.schemas.ts
export type ContactMethodEnum = typeof ContactMethodEnum[keyof typeof ContactMethodEnum];


export const ContactMethodEnum = {
  email: 'email',
  phone: 'phone',
  both: 'both',
} as const;